update requirements and basic websocket
